refactor: 채팅 서비스 CQRS 패턴 적용 (#22)
