+++ src/main/java/com/hyunmin/gpt/domain/chat/controller/ChatController.java
@@ -3,8 +3,9 @@ package com.hyunmin.gpt.domain.chat.controller;
 import com.hyunmin.gpt.domain.chat.dto.ChatGptRequestDto;
 import com.hyunmin.gpt.domain.chat.dto.ChatRequestDto;
 import com.hyunmin.gpt.domain.chat.dto.ChatResponseDto;
+import com.hyunmin.gpt.domain.chat.service.ChatCommandService;
 import com.hyunmin.gpt.domain.chat.service.ChatGptService;
-import com.hyunmin.gpt.domain.chat.service.ChatService;
+import com.hyunmin.gpt.domain.chat.service.ChatQueryService;
 import com.hyunmin.gpt.domain.message.service.MessageService;
 import com.hyunmin.gpt.global.security.annotation.AuthMember;
 import jakarta.validation.Valid;
@@ -22,13 +23,14 @@ import reactor.core.publisher.Flux;
 @RequestMapping("/api/v1/chat")
 public class ChatController {
 
-    private final ChatService chatService;
+    private final ChatQueryService chatQueryService;
+    private final ChatCommandService chatCommandService;
     private final ChatGptService chatGptService;
     private final MessageService messageService;
 
     @PostMapping(produces = MediaType.TEXT_EVENT_STREAM_VALUE)
     public ResponseEntity<Flux<String>> streamChat(@AuthMember Long memberId, @RequestBody @Valid ChatRequestDto requestDto) {
-        String chatId = chatService.getOrCreateChatId(memberId, requestDto);
+        String chatId = chatCommandService.getOrCreateChatId(memberId, requestDto);
         ChatGptRequestDto gptRequestDto = messageService.readMessagesForChatGpt(chatId, requestDto.content());
         Flux<String> responseFlux = chatGptService.streamChat(chatId, gptRequestDto, requestDto.content());
         return ResponseEntity.ok()
@@ -38,7 +40,7 @@ public class ChatController {
 
     @GetMapping
     public ResponseEntity<Slice<ChatResponseDto>> readChats(@AuthMember Long memberId, @ParameterObject Pageable pageable) {
-        Slice<ChatResponseDto> responseDtoSlice = chatService.readChats(memberId, pageable);
+        Slice<ChatResponseDto> responseDtoSlice = chatQueryService.readChats(memberId, pageable);
         return ResponseEntity.ok(responseDtoSlice);
     }
 }
