+++ src/main/java/com/hyunmin/gpt/domain/chat/service/ChatCommandService.java
@@ -1,0 +1,39 @@
+package com.hyunmin.gpt.domain.chat.service;
+
+import com.hyunmin.gpt.domain.chat.dto.ChatRequestDto;
+import com.hyunmin.gpt.domain.chat.dto.ChatResponseDto;
+import com.hyunmin.gpt.domain.chat.entity.Chat;
+import com.hyunmin.gpt.domain.chat.repository.ChatRepository;
+import com.hyunmin.gpt.global.common.entity.Member;
+import com.hyunmin.gpt.global.common.repository.MemberRepository;
+import com.hyunmin.gpt.global.exception.GeneralException;
+import com.hyunmin.gpt.global.exception.code.ErrorCode;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class ChatCommandService {
+
+    private final ChatQueryService chatQueryService;
+    private final ChatRepository chatRepository;
+    private final MemberRepository memberRepository;
+
+    public ChatResponseDto createChat(Long memberId, ChatRequestDto requestDto) {
+        Chat chat = requestDto.toEntity();
+        if (memberId != null) {
+            Member member = memberRepository.findById(memberId)
+                    .orElseThrow(() -> new GeneralException(ErrorCode.ACCOUNT_NOT_FOUND));
+            chat.setMember(member);
+        }
+        return ChatResponseDto.from(chatRepository.save(chat));
+    }
+
+    public String getOrCreateChatId(Long memberId, ChatRequestDto requestDto) {
+        return requestDto.chatId() != null ?
+                chatQueryService.readChat(memberId, requestDto.chatId()).id() :
+                createChat(memberId, requestDto).id();
+    }
+}
